requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
beanie==1.26.0
motor==3.4.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
firebase-admin==6.5.0
qrcode==7.4.2
Pillow==10.4.0
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
